requirements.txt
streamlit>=1.28.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
spacy>=3.7.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0